+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-specific)/(stage-1)/all-supervisors/_components/all-supervisors-columns.tsx
@@ -35,7 +35,7 @@ import {
 } from "@/lib/utils/permissions/stage-check";
 
 import { Role, Stage } from "@/db/types";
-import { SupervisorDTO } from "@/dto/user/supervisor";
+import { SupervisorDTO } from "@/dto";
 
 export function useAllSupervisorsColumns({
   roles,
