+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-specific)/(stage-1)/all-supervisors/_components/all-supervisors-data-table.tsx
@@ -11,7 +11,7 @@ import { useAllSupervisorsColumns } from "./all-supervisors-columns";
 
 import { spacesLabels } from "@/config/spaces";
 import { Role } from "@/db/types";
-import { SupervisorDTO } from "@/dto/user/supervisor";
+import { SupervisorDTO } from "@/dto";
 
 export function SupervisorsDataTable({
   roles,
