+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-specific)/(stage-4)/algorithms/_components/algorithm-columns.tsx
@@ -21,7 +21,7 @@ import {
 } from "@/components/yes-no-action";
 
 import { api } from "@/lib/trpc/client";
-import { AlgorithmDTO } from "@/dto/algorithm";
+import { AlgorithmDTO } from "@/dto";
 
 import { useAlgorithmUtils } from "./algorithm-context";
 import { RunAlgorithmButton } from "./run-algorithm-button";
