+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-specific)/(stage-4)/algorithms/_components/run-algorithm-button.tsx
@@ -4,7 +4,7 @@ import { useInstanceParams } from "@/components/params-context";
 import { Button } from "@/components/ui/button";
 
 import { api } from "@/lib/trpc/client";
-import { AlgorithmDTO } from "@/dto/algorithm";
+import { AlgorithmDTO } from "@/dto";
 
 import { useAlgorithmUtils } from "./algorithm-context";
