+++ src/server/routers/institution/instance/algorithm/_utils/execute-matching-algorithm.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 import { builtInAlgSchema } from "@/lib/validations/algorithm";
-import { AlgorithmDTO } from "@/dto/algorithm";
+import { AlgorithmDTO } from "@/dto";
 import {
   MatchingDataDTO,
   MatchingDataWithArgs,
@@ -9,7 +9,7 @@ import {
 } from "@/lib/validations/matching";
 
 import { AlgorithmFlag } from "@/db/types";
-import { AlgorithmRunResult } from "@/dto/result/algorithm-run-result";
+import { AlgorithmRunResult } from "@/dto";
 import { env } from "@/env";
 
 // ? where should this live
