+++ src/server/routers/institution/instance/algorithm/_utils/get-algorithms-in-order.ts
@@ -2,7 +2,7 @@ import {
   builtInAlgorithms,
   builtInAlgSchema,
 } from "@/lib/validations/algorithm";
-import { AlgorithmDTO } from "@/dto/algorithm";
+import { AlgorithmDTO } from "@/dto";
 
 // should be displayed in the order they were created, except the built-in ones which should be displayed first
 /**
